Added support for pre-persist Event hooks
This allows the modification of an event before it is persisted to the database. Hooks can be enabled/disabled based on account/group (and other parameters) and adjust the values of the event before insertion. The hook added in this PR is enabled for ACMB and Cygnus groups, and ensures that all incoming customer ids are set as MongoIds. This was needed because Olytics customer cookies still exist with legacy data, such as Omeda encrypted customer ids and string Merrick user ids. This hook converts these values (where possible) into Mongo ObjectId() format.